+++ BraintreeCore/BTAPIClient.m
@@ -9,7 +9,9 @@
 NSString *const BTAPIClientErrorDomain = @"com.braintreepayments.BTAPIClientErrorDomain";
 
 @interface BTAPIClient ()
+
 @property (nonatomic, strong) dispatch_queue_t configurationQueue;
+
 @end
 
 @implementation BTAPIClient
@@ -26,44 +28,64 @@ NSString *const BTAPIClientErrorDomain = @"com.braintreepayments.BTAPIClientErro
     }
 
     if (self = [super init]) {
+        BTAPIClientAuthorizationType authorizationType = [[self class] authorizationTypeForAuthorization:authorization];
+        switch (authorizationType) {
+            case BTAPIClientAuthorizationTypeTokenizationKey: {
+                NSURL *baseURL = [BTAPIClient baseURLFromTokenizationKey:authorization];
+
+                if (!baseURL) {
+                    NSString *reason = @"BTClient could not initialize because the provided tokenization key was invalid";
+                    [[BTLogger sharedLogger] error:reason];
+                    return nil;
+                }
+
+                _tokenizationKey = authorization;
+                _configurationHTTP = [[BTHTTP alloc] initWithBaseURL:baseURL tokenizationKey:authorization];
+
+                if (sendAnalyticsEvent) {
+                    [self queueAnalyticsEvent:@"ios.started.client-key"];
+                }
+                break;
+            }
+
+            case BTAPIClientAuthorizationTypeClientToken: {
+                NSError *error;
+                _clientToken = [[BTClientToken alloc] initWithClientToken:authorization error:&error];
+                if (error) { [[BTLogger sharedLogger] error:[error localizedDescription]]; }
+                if (!_clientToken) {
+                    [[BTLogger sharedLogger] error:@"BTClient could not initialize because the provided clientToken was invalid"];
+                    return nil;
+                }
+
+                _configurationHTTP = [[BTHTTP alloc] initWithClientToken:self.clientToken];
+
+                if (sendAnalyticsEvent) {
+                    [self queueAnalyticsEvent:@"ios.started.client-token"];
+                }
+                break;
+            }
+
+            case BTAPIClientAuthorizationTypePayPalUAT: {
+                NSError *error;
+                _payPalUAT = [[BTPayPalUAT alloc] initWithUATString:authorization error:&error];
+                if (!_payPalUAT || error) {
+                    [[BTLogger sharedLogger] error:@"BTClient could not initialize because the provided PayPal UAT was invalid"];
+                    [[BTLogger sharedLogger] error:[error localizedDescription]];
+                    return nil;
+                }
+                
+                _configurationHTTP = [[BTHTTP alloc] initWithPayPalUAT:_payPalUAT];
+
+                if (sendAnalyticsEvent) {
+                    [self queueAnalyticsEvent:@"ios.started.paypal-uat"];
+                }
+                break;
+            }
+        }
+
         _metadata = [[BTClientMetadata alloc] init];
         _configurationQueue = dispatch_queue_create("com.braintreepayments.BTAPIClient", DISPATCH_QUEUE_SERIAL);
 
-        NSRegularExpression *isTokenizationKeyRegExp = [NSRegularExpression regularExpressionWithPattern:@"^[a-zA-Z0-9]+_[a-zA-Z0-9]+_[a-zA-Z0-9_]+$" options:0 error:NULL];
-        NSTextCheckingResult *tokenizationKeyMatch = [isTokenizationKeyRegExp firstMatchInString:authorization options:0 range: NSMakeRange(0, authorization.length)];
-
-        if (tokenizationKeyMatch) {
-            NSURL *baseURL = [BTAPIClient baseURLFromTokenizationKey:authorization];
-
-            if (!baseURL) {
-                NSString *reason = @"BTClient could not initialize because the provided tokenization key was invalid";
-                [[BTLogger sharedLogger] error:reason];
-                return nil;
-            }
-
-            _tokenizationKey = authorization;
-            _configurationHTTP = [[BTHTTP alloc] initWithBaseURL:baseURL tokenizationKey:authorization];
-            
-            if (sendAnalyticsEvent) {
-                [self queueAnalyticsEvent:@"ios.started.client-key"];
-            }
-        } else {
-            NSError *error;
-            _clientToken = [[BTClientToken alloc] initWithClientToken:authorization error:&error];
-            if (error) { [[BTLogger sharedLogger] error:[error localizedDescription]]; }
-            if (!_clientToken) {
-                NSString *reason = @"BTClient could not initialize because the provided clientToken was invalid";
-                [[BTLogger sharedLogger] error:reason];
-                return nil;
-            }
-
-            _configurationHTTP = [[BTHTTP alloc] initWithClientToken:self.clientToken];
-
-            if (sendAnalyticsEvent) {
-                [self queueAnalyticsEvent:@"ios.started.client-token"];
-            }
-        }
-        
         // BTHTTP's default NSURLSession does not cache responses, but we want the BTHTTP instance that fetches configuration to cache aggressively
         NSURLSessionConfiguration *configuration = [NSURLSessionConfiguration defaultSessionConfiguration];
         static NSURLCache *configurationCache;
@@ -80,8 +102,24 @@ NSString *const BTAPIClientErrorDomain = @"com.braintreepayments.BTAPIClientErro
             //noop
         }];
     }
-    
+
     return self;
+}
+
++ (BTAPIClientAuthorizationType)authorizationTypeForAuthorization:(NSString *)authorization {
+    NSRegularExpression *isTokenizationKeyRegExp = [NSRegularExpression regularExpressionWithPattern:@"^[a-zA-Z0-9]+_[a-zA-Z0-9]+_[a-zA-Z0-9_]+$" options:0 error:NULL];
+    NSTextCheckingResult *tokenizationKeyMatch = [isTokenizationKeyRegExp firstMatchInString:authorization options:0 range: NSMakeRange(0, authorization.length)];
+
+    NSRegularExpression *isPayPalUATRegExp = [NSRegularExpression regularExpressionWithPattern:@"^[a-zA-Z0-9]+\\.[a-zA-Z0-9]+\\.[a-zA-Z0-9_-]+$" options:0 error:NULL];
+    NSTextCheckingResult *payPalUATMatch = [isPayPalUATRegExp firstMatchInString:authorization options:0 range: NSMakeRange(0, authorization.length)];
+
+    if (tokenizationKeyMatch) {
+        return BTAPIClientAuthorizationTypeTokenizationKey;
+    } else if (payPalUATMatch) {
+        return BTAPIClientAuthorizationTypePayPalUAT;
+    } else {
+        return BTAPIClientAuthorizationTypeClientToken;
+    }
 }
 
 - (instancetype)copyWithSource:(BTClientMetadataSourceType)source
@@ -93,6 +131,8 @@ NSString *const BTAPIClientErrorDomain = @"com.braintreepayments.BTAPIClientErro
         copiedClient = [[[self class] alloc] initWithAuthorization:self.clientToken.originalValue sendAnalyticsEvent:NO];
     } else if (self.tokenizationKey) {
         copiedClient = [[[self class] alloc] initWithAuthorization:self.tokenizationKey sendAnalyticsEvent:NO];
+    } else if (self.payPalUAT) {
+        copiedClient = [[[self class] alloc] initWithAuthorization:self.payPalUAT.token sendAnalyticsEvent:NO];
     } else {
         NSAssert(NO, @"Cannot copy an API client that does not specify a client token or tokenization key");
     }
@@ -258,10 +298,15 @@ NSString *const BTAPIClientErrorDomain = @"com.braintreepayments.BTAPIClientErro
     //       queue is a background queue to guarantee atomic access to the remote configuration resource.
     dispatch_async(self.configurationQueue, ^{
         __block NSError *fetchError;
-        
+
         dispatch_semaphore_t semaphore = dispatch_semaphore_create(0);
         __block BTConfiguration *configuration;
-        NSString *configPath = self.tokenizationKey ? @"v1/configuration" : [self.clientToken.configURL absoluteString];
+        NSString *configPath = @"v1/configuration"; // Default for tokenizationKey
+        if (self.clientToken) {
+            configPath = [self.clientToken.configURL absoluteString];
+        } else if (self.payPalUAT) {
+            configPath = [self.payPalUAT.configURL absoluteString];
+        }
         [self.configurationHTTP GET:configPath parameters:@{ @"configVersion": @"3" } completion:^(BTJSON * _Nullable body, NSHTTPURLResponse * _Nullable response, NSError * _Nullable error) {
             if (error) {
                 fetchError = error;
@@ -286,6 +331,8 @@ NSString *const BTAPIClientErrorDomain = @"com.braintreepayments.BTAPIClientErro
                         self.http = [[BTHTTP alloc] initWithBaseURL:baseURL authorizationFingerprint:self.clientToken.authorizationFingerprint];
                     } else if (self.tokenizationKey) {
                         self.http = [[BTHTTP alloc] initWithBaseURL:baseURL tokenizationKey:self.tokenizationKey];
+                    } else if (self.payPalUAT) {
+                        self.http = [[BTHTTP alloc] initWithBaseURL:baseURL authorizationFingerprint:self.payPalUAT.token];
                     }
                 }
                 if (!self.graphQL) {
@@ -294,14 +341,16 @@ NSString *const BTAPIClientErrorDomain = @"com.braintreepayments.BTAPIClientErro
                         self.graphQL = [[BTGraphQLHTTP alloc] initWithBaseURL:graphQLBaseURL authorizationFingerprint:self.clientToken.authorizationFingerprint];
                     } else if (self.tokenizationKey) {
                         self.graphQL = [[BTGraphQLHTTP alloc] initWithBaseURL:graphQLBaseURL tokenizationKey:self.tokenizationKey];
+                    } else if (self.payPalUAT) {
+                        self.graphQL = [[BTGraphQLHTTP alloc] initWithBaseURL:graphQLBaseURL authorizationFingerprint:self.payPalUAT.token];
                     }
                 }
             }
-            
+
             // Important: Unlock semaphore in all cases
             dispatch_semaphore_signal(semaphore);
         }];
-        
+
         dispatch_semaphore_wait(semaphore, DISPATCH_TIME_FOREVER);
 
         dispatch_async(dispatch_get_main_queue(), ^{
@@ -321,10 +370,10 @@ NSString *const BTAPIClientErrorDomain = @"com.braintreepayments.BTAPIClientErro
         analyticsService = [[BTAnalyticsService alloc] initWithAPIClient:self];
         analyticsService.flushThreshold = 5;
     });
-    
+
     // The analytics service may be overridden by unit tests. In that case, return the ivar and not the singleton
     if (_analyticsService) return _analyticsService;
-    
+
     return analyticsService;
 }
 
@@ -414,7 +463,7 @@ NSString *const BTAPIClientErrorDomain = @"com.braintreepayments.BTAPIClientErro
     if (self.http && self.http.session) {
         [self.http.session finishTasksAndInvalidate];
     }
-    
+
     if (self.braintreeAPI && self.braintreeAPI.session) {
         [self.braintreeAPI.session finishTasksAndInvalidate];
     }