+++ BraintreeCore/BTHTTP.m
@@ -7,6 +7,7 @@
 #import "BTAPIPinnedCertificates.h"
 #import "BTURLUtils.h"
 #import "BTLogger_Internal.h"
+#import "BTPayPalUAT.h"
 
 @interface BTHTTP () <NSURLSessionDelegate>
 
@@ -66,6 +67,10 @@
 
 - (instancetype)initWithClientToken:(BTClientToken *)clientToken {
     return [self initWithBaseURL:[clientToken.json[@"clientApiUrl"] asURL] authorizationFingerprint:clientToken.authorizationFingerprint];
+}
+
+- (instancetype)initWithPayPalUAT:(BTPayPalUAT *)payPalUAT {
+    return [self initWithBaseURL:payPalUAT.baseBraintreeURL authorizationFingerprint:payPalUAT.token];
 }
 
 - (instancetype)copyWithZone:(NSZone *)zone {