+++ UnitTests/BraintreeCore_UnitTests/BTAPIClient_SwiftTests.swift
@@ -1,22 +1,22 @@
 import XCTest
 
 class BTAPIClient_SwiftTests: XCTestCase {
-    
+
     // MARK: - Initialization
-    
+
     func testAPIClientInitialization_withValidTokenizationKey_returnsClientWithTokenizationKey() {
         let apiClient = BTAPIClient(authorization: "development_testing_integration_merchant_id")!
         XCTAssertEqual(apiClient.tokenizationKey, "development_testing_integration_merchant_id")
     }
-    
-    func testAPIClientInitialization_withInvalidTokenizationKey_returnsNil() {
+
+    func testAPIClientInitialization_withInvalidAuthorization_returnsNil() {
         XCTAssertNil(BTAPIClient(authorization: "invalid"))
     }
-    
-    func testAPIClientInitialization_withEmptyTokenizationKey_returnsNil() {
+
+    func testAPIClientInitialization_withEmptyAuthorization_returnsNil() {
         XCTAssertNil(BTAPIClient(authorization: ""))
     }
-    
+
     func testAPIClientInitialization_withValidClientToken_returnsClientWithClientToken() {
         let clientToken = BTTestClientTokenFactory.token(withVersion: 2)
         let apiClient = BTAPIClient(authorization: clientToken!)
@@ -34,6 +34,38 @@ class BTAPIClient_SwiftTests: XCTestCase {
         self.measure() {
             _ = BTAPIClient(authorization: clientToken!)
         }
+    }
+
+    func testAPIClientInitialization_withValidPayPalUAT_returnsClientWithPayPalUAT() {
+        let payPalUAT = "123.ewogICAiaXNzIjoiaHR0cHM6Ly9hcGkucGF5cGFsLmNvbSIsCiAgICJzdWIiOiJQYXlQYWw6ZmFrZS1wcC1tZXJjaGFudCIsCiAgICJhY3IiOlsKICAgICAgImNsaWVudCIKICAgXSwKICAgInNjb3BlcyI6WwogICAgICAiQnJhaW50cmVlOlZhdWx0IgogICBdLAogICAiZXhwIjoxNTcxOTgwNTA2LAogICAiZXh0ZXJuYWxfaWRzIjpbCiAgICAgICJQYXlQYWw6ZmFrZS1wcC1tZXJjaGFudCIsCiAgICAgICJCcmFpbnRyZWU6ZmFrZS1idC1tZXJjaGFudCIKICAgXSwKICAgImp0aSI6ImZha2UtanRpIgp9.456"
+        let apiClient = BTAPIClient(authorization: payPalUAT)
+        XCTAssertEqual(apiClient?.payPalUAT?.token, payPalUAT)
+    }
+
+    func testAPIClientIntialization_withInvalidPayPalUAT_returnsNil() {
+        let payPalUAT = "broken.paypal.uat"
+        let apiClient = BTAPIClient(authorization: payPalUAT)
+        XCTAssertNil(apiClient)
+    }
+
+    // MARK: - authorizationType
+    
+    func testAPIClientAuthorizationType_forTokenizationKey() {
+        let tokenizationKey = "sandbox_test1xxx_123xx2swdz6nxxx7"
+        let apiClientAuthType = BTAPIClient.authorizationType(forAuthorization: tokenizationKey)
+        XCTAssertEqual(apiClientAuthType, .tokenizationKey)
+    }
+
+    func testAPIClientAuthorizationType_forClientToken() {
+        let clientToken = "1234abc=="
+        let apiClientAuthType = BTAPIClient.authorizationType(forAuthorization: clientToken)
+        XCTAssertEqual(apiClientAuthType, .clientToken)
+    }
+
+    func testAPIClientAuthorizationType_forPayPalUAT() {
+        let payPalUAT = "1a.2b.3c-_"
+        let apiClientAuthType = BTAPIClient.authorizationType(forAuthorization: payPalUAT)
+        XCTAssertEqual(apiClientAuthType, .payPalUAT)
     }
 
     // MARK: - Copy
@@ -65,22 +97,22 @@ class BTAPIClient_SwiftTests: XCTestCase {
         let copiedApiClient = apiClient?.copy(with: .payPalBrowser, integration: .dropIn)
         XCTAssertTrue(copiedApiClient !== apiClient)
     }
-    
+
     // MARK: - fetchOrReturnRemoteConfiguration
-    
+
     func testFetchOrReturnRemoteConfiguration_performsGETWithCorrectPayload() {
         let apiClient = BTAPIClient(authorization: "development_testing_integration_merchant_id", sendAnalyticsEvent: false)!
         let mockHTTP = BTFakeHTTP()!
         mockHTTP.stubRequest("GET", toEndpoint: "/v1/configuration", respondWith: [], statusCode: 200)
         apiClient.configurationHTTP = mockHTTP
-       
+
         let expectation = self.expectation(description: "Callback invoked")
         apiClient.fetchOrReturnRemoteConfiguration() { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/configuration")
             XCTAssertEqual(mockHTTP.lastRequestParameters?["configVersion"] as? String, "3")
             expectation.fulfill()
         }
-        
+
         waitForExpectations(timeout: 1, handler: nil)
     }
 
@@ -91,7 +123,7 @@ class BTAPIClient_SwiftTests: XCTestCase {
         let mockHTTP = BTFakeHTTP()!
         mockHTTP.stubRequest("GET", toEndpoint: "/client_api/v1/payment_methods", respondWith: [], statusCode: 200)
         apiClient.http = mockHTTP
-       
+
         var expectation = self.expectation(description: "Callback invoked")
         apiClient.fetchPaymentMethodNonces() { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/payment_methods")
@@ -99,28 +131,28 @@ class BTAPIClient_SwiftTests: XCTestCase {
             XCTAssertEqual(mockHTTP.lastRequestParameters!["session_id"] as? String, apiClient.metadata.sessionId)
             expectation.fulfill()
         }
-        
+
         waitForExpectations(timeout: 1, handler: nil)
-       
+
         expectation = self.expectation(description: "Callback invoked")
         apiClient.fetchPaymentMethodNonces(true) { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/payment_methods")
             XCTAssertEqual(mockHTTP.lastRequestParameters!["default_first"] as? String, "true")
             expectation.fulfill()
         }
-        
+
         waitForExpectations(timeout: 1, handler: nil)
-        
+
         expectation = self.expectation(description: "Callback invoked")
         apiClient.fetchPaymentMethodNonces(false) { _,_  in
             XCTAssertEqual(mockHTTP.lastRequestEndpoint, "v1/payment_methods")
             XCTAssertEqual(mockHTTP.lastRequestParameters!["default_first"] as? String, "false")
             expectation.fulfill()
         }
-        
+
         waitForExpectations(timeout: 1, handler: nil)
     }
-    
+
     func testFetchPaymentMethods_returnsPaymentMethodNonces() {
         let apiClient = BTAPIClient(authorization: BTValidTestClientToken, sendAnalyticsEvent: false)!
         let stubHTTP = BTFakeHTTP()!
@@ -146,17 +178,17 @@ class BTAPIClient_SwiftTests: XCTestCase {
             ] ]
         stubHTTP.stubRequest("GET", toEndpoint: "/client_api/v1/payment_methods", respondWith: stubbedResponse, statusCode: 200)
         apiClient.http = stubHTTP
-       
+
         let expectation = self.expectation(description: "Callback invoked")
         apiClient.fetchPaymentMethodNonces() { (paymentMethodNonces, error) in
             guard let paymentMethodNonces = paymentMethodNonces else {
                 XCTFail()
                 return
             }
-            
+
             XCTAssertNil(error)
             XCTAssertEqual(paymentMethodNonces.count, 2)
-            
+
             guard let cardNonce = paymentMethodNonces[0] as? BTCardNonce else {
                 XCTFail()
                 return
@@ -165,17 +197,17 @@ class BTAPIClient_SwiftTests: XCTestCase {
                 XCTFail()
                 return
             }
-            
+
             XCTAssertEqual(cardNonce.nonce, "fake-nonce")
             XCTAssertEqual(cardNonce.localizedDescription, "ending in 05")
             XCTAssertEqual(cardNonce.lastTwo, "05")
             XCTAssertTrue(cardNonce.cardNetwork == BTCardNetwork.AMEX)
             XCTAssertTrue(cardNonce.isDefault)
-            
+
             XCTAssertEqual(paypalNonce.nonce, "fake-nonce")
             XCTAssertEqual(paypalNonce.localizedDescription, "jane.doe@example.com")
             XCTAssertFalse(paypalNonce.isDefault)
-            
+
             expectation.fulfill()
         }
 
@@ -241,7 +273,7 @@ class BTAPIClient_SwiftTests: XCTestCase {
     }
 
     // MARK: - Analytics
-    
+
     func testAnalyticsService_byDefault_isASingleton() {
         let firstAPIClient = BTAPIClient(authorization: "development_testing_integration_merchant_id")!
         let secondAPIClient = BTAPIClient(authorization: "development_testing_integration_merchant_id")!