+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add support for iOS 13 SceneDelegate to `BTAppSwitch`
 * Add `lastFour` property to `BTCardNonce`
 * Make `BTURLUtils.h` public
+* Add support for authorizing the Braintree SDK with a `PayPalUAT` (universal access token)
 * Remove `AddressBook.framework` from Podspec (thanks @ignotusverum)
 
 ## 4.32.1 (2020-02-21)